+++ CHANGELOG.md
@@ -56,3 +56,6 @@ Initial version
 * xmlValidate now returns value (string) if execution fails
 * CLI
   * return exit status 1 if command failed
+* xml config
+  * added support for dataTraceConf
+  * added support for cpuSpeed element in serial config (required for swo mode in serial tracing)